feature: added memory information, splitted pkginfo file from sysinfo
